split the test
