+++ test/srt/test_vlm_accuracy.py
@@ -10,15 +10,8 @@ import requests
 import torch
 import torch.nn.functional as F
 from PIL import Image
-from transformers import (
-    AutoModel,
-    AutoProcessor,
-    AutoTokenizer,
-    Gemma3ForConditionalGeneration,
-    Qwen2_5_VLForConditionalGeneration,
-)
+from transformers import AutoModel, AutoProcessor, AutoTokenizer
 
-from sglang import Engine
 from sglang.srt.configs.model_config import ModelConfig
 from sglang.srt.conversation import generate_chat_conv
 from sglang.srt.managers.mm_utils import embed_mm_inputs
@@ -243,130 +236,3 @@ class TestMiniCPMVLogits(VisionLLMLogitsBase):
             )
 
         self.compare_outputs(sglang_output, hf_output)
-
-
-class VisionLLMTestBase(VisionLLMLogitsBase):
-    model_path = None
-    chat_template = None
-    processor = None
-    visual = None  # Should be a callable for precomputed features
-
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        assert cls.model_path is not None, "Set model_path in subclass"
-        assert cls.chat_template is not None, "Set chat_template in subclass"
-        cls.processor = AutoProcessor.from_pretrained(
-            cls.model_path, trust_remote_code=True, use_fast=True
-        )
-        cls._init_visual()
-
-    @classmethod
-    def _init_visual(cls):
-        """Override in subclass to set up cls.visual as a callable for precomputed features."""
-        raise NotImplementedError
-
-    def setUp(self):
-        self.engine = Engine(
-            model_path=self.model_path,
-            chat_template=self.chat_template,
-            device=self.device.type,
-            mem_fraction_static=0.8,
-            enable_multimodal=True,
-            disable_cuda_graph=True,
-        )
-
-    def tearDown(self):
-        self.engine.shutdown()
-
-    async def test_understands_image(self):
-        req = self.get_completion_request()
-        conv = generate_chat_conv(req, template_name=self.chat_template)
-        text = conv.get_prompt()
-        output = await self.engine.async_generate(
-            prompt=text,
-            image_data=[self.main_image],
-            sampling_params=dict(temperature=0.0),
-        )
-        self.assertIn("taxi", output["text"].lower())
-
-    async def test_understands_precomputed_features(self):
-        req = self.get_completion_request()
-        processor_output = self.get_processor_output(req=req)
-        with torch.inference_mode():
-            precomputed_features = self.__class__.visual(processor_output)
-        output = await self.engine.async_generate(
-            input_ids=processor_output["input_ids"][0].detach().cpu().tolist(),
-            image_data=[
-                self._precomputed_image_data(processor_output, precomputed_features)
-            ],
-            sampling_params=dict(temperature=0.0),
-        )
-        self.assertIn("taxi", output["text"].lower())
-
-    async def test_understands_pixel_values(self):
-        req = self.get_completion_request()
-        processor_output = self.get_processor_output(req=req)
-        output = await self.engine.async_generate(
-            input_ids=processor_output["input_ids"][0].detach().cpu().tolist(),
-            image_data=[self._pixel_values_image_data(processor_output)],
-            sampling_params=dict(temperature=0.0),
-        )
-        self.assertIn("taxi", output["text"].lower())
-
-    def _precomputed_image_data(self, processor_output, precomputed_features):
-        """Override in subclass if needed."""
-        return dict(
-            modality="IMAGE",
-            precomputed_features=precomputed_features,
-        )
-
-    def _pixel_values_image_data(self, processor_output):
-        """Override in subclass if needed."""
-        return dict(
-            modality="IMAGE",
-            pixel_values=processor_output["pixel_values"][0],
-        )
-
-
-class TestQwenVLUnderstandsImage(VisionLLMTestBase):
-    model_path = "Qwen/Qwen2.5-VL-3B-Instruct"
-    chat_template = "qwen2-vl"
-
-    @classmethod
-    def _init_visual(cls):
-        cls.visual_model = (
-            Qwen2_5_VLForConditionalGeneration.from_pretrained(
-                cls.model_path, torch_dtype=torch.bfloat16
-            )
-            .eval()
-            .visual.to(cls.device)
-        )
-        cls.visual = lambda processor_output: cls.visual_model(
-            processor_output["pixel_values"], processor_output["image_grid_thw"]
-        )
-
-    def _pixel_values_image_data(self, processor_output):
-        return dict(
-            modality="IMAGE",
-            image_grid_thws=processor_output["image_grid_thw"],
-            pixel_values=processor_output["pixel_values"],
-        )
-
-
-class TestGemmaUnderstandsImage(VisionLLMTestBase):
-    model_path = "google/gemma-3-4b-it"
-    chat_template = "gemma-it"
-
-    @classmethod
-    def _init_visual(cls):
-        model = Gemma3ForConditionalGeneration.from_pretrained(
-            cls.model_path, torch_dtype=torch.bfloat16
-        )
-        cls.vision_tower = model.vision_tower.eval().to(cls.device)
-        cls.mm_projector = model.multi_modal_projector.eval().to(cls.device)
-        cls.visual = lambda processor_output: cls.mm_projector(
-            cls.vision_tower(
-                pixel_values=processor_output["pixel_values"]
-            ).last_hidden_state
-        )
